refactoring, move the logic from the command file to the main aspect file
