+++ scopes/mcp/cli-mcp-server/cli-mcp-server.main.runtime.ts
@@ -3,6 +3,9 @@
 
 import { CLIAspect, CLIMain, Command, getArgsData, getCommandName, getFlagsData, MainRuntime } from '@teambit/cli';
 import childProcess from 'child_process';
+import fs from 'fs-extra';
+import path from 'path';
+import { homedir } from 'os';
 import { CliMcpServerAspect } from './cli-mcp-server.aspect';
 import { McpServerCmd, McpStartCmd } from './mcp-server.cmd';
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
@@ -1078,6 +1081,250 @@ export class CliMcpServerMain {
       } else {
         throw error;
       }
+    }
+  }
+
+  // Setup command business logic methods
+  getEditorDisplayName(editor: string): string {
+    switch (editor) {
+      case 'vscode':
+        return 'VS Code';
+      case 'cursor':
+        return 'Cursor';
+      case 'windsurf':
+        return 'Windsurf';
+      default:
+        return editor;
+    }
+  }
+
+  buildMcpServerArgs(options: {
+    extended?: boolean;
+    consumerProject?: boolean;
+    includeOnly?: string;
+    includeAdditional?: string;
+    exclude?: string;
+  }): string[] {
+    const { extended, consumerProject, includeOnly, includeAdditional, exclude } = options;
+    const args = ['mcp-server'];
+
+    if (extended) {
+      args.push('--extended');
+    }
+
+    if (consumerProject) {
+      args.push('--consumer-project');
+    }
+
+    if (includeOnly) {
+      args.push('--include-only', includeOnly);
+    }
+
+    if (includeAdditional) {
+      args.push('--include-additional', includeAdditional);
+    }
+
+    if (exclude) {
+      args.push('--exclude', exclude);
+    }
+
+    return args;
+  }
+
+  async readJsonFile(filePath: string): Promise<any> {
+    if (!(await fs.pathExists(filePath))) {
+      return {};
+    }
+
+    try {
+      const content = await fs.readFile(filePath, 'utf8');
+      return JSON.parse(content);
+    } catch (error) {
+      throw new Error(`Failed to parse ${path.basename(filePath)}: ${(error as Error).message}`);
+    }
+  }
+
+  async setupVSCode(options: {
+    extended?: boolean;
+    consumerProject?: boolean;
+    includeOnly?: string;
+    includeAdditional?: string;
+    exclude?: string;
+    isGlobal: boolean;
+  }): Promise<void> {
+    const { isGlobal } = options;
+
+    // Determine settings.json path
+    const settingsPath = this.getVSCodeSettingsPath(isGlobal);
+
+    // Ensure directory exists
+    await fs.ensureDir(path.dirname(settingsPath));
+
+    // Read existing settings or create empty object
+    const settings = await this.readJsonFile(settingsPath);
+
+    // Build MCP server args
+    const args = this.buildMcpServerArgs(options);
+
+    // Create or update MCP configuration
+    if (!settings.mcp) {
+      settings.mcp = {};
+    }
+
+    if (!settings.mcp.servers) {
+      settings.mcp.servers = {};
+    }
+
+    settings.mcp.servers['bit-cli'] = {
+      command: 'bit',
+      args: args,
+    };
+
+    // Write updated settings
+    await fs.writeFile(settingsPath, JSON.stringify(settings, null, 2));
+  }
+
+  getVSCodeSettingsPath(isGlobal: boolean): string {
+    if (isGlobal) {
+      // Global VS Code settings
+      const platform = process.platform;
+      switch (platform) {
+        case 'win32':
+          return path.join(homedir(), 'AppData', 'Roaming', 'Code', 'User', 'settings.json');
+        case 'darwin':
+          return path.join(homedir(), 'Library', 'Application Support', 'Code', 'User', 'settings.json');
+        case 'linux':
+          return path.join(homedir(), '.config', 'Code', 'User', 'settings.json');
+        default:
+          throw new Error(`Unsupported platform: ${platform}`);
+      }
+    } else {
+      // Workspace-specific settings
+      const workspaceDir = process.cwd();
+      return path.join(workspaceDir, '.vscode', 'settings.json');
+    }
+  }
+
+  async setupCursor(options: {
+    extended?: boolean;
+    consumerProject?: boolean;
+    includeOnly?: string;
+    includeAdditional?: string;
+    exclude?: string;
+    isGlobal: boolean;
+  }): Promise<void> {
+    const { isGlobal } = options;
+
+    // Determine mcp.json path
+    const mcpConfigPath = this.getCursorSettingsPath(isGlobal);
+
+    // Ensure directory exists
+    await fs.ensureDir(path.dirname(mcpConfigPath));
+
+    // Read existing MCP configuration or create empty object
+    const mcpConfig = await this.readJsonFile(mcpConfigPath);
+
+    // Build MCP server args
+    const args = this.buildMcpServerArgs(options);
+
+    // Create or update MCP configuration for Cursor
+    if (!mcpConfig.mcpServers) {
+      mcpConfig.mcpServers = {};
+    }
+
+    mcpConfig.mcpServers.bit = {
+      type: 'stdio',
+      command: 'bit',
+      args: args,
+    };
+
+    // Write updated MCP configuration
+    await fs.writeFile(mcpConfigPath, JSON.stringify(mcpConfig, null, 2));
+  }
+
+  getCursorSettingsPath(isGlobal: boolean): string {
+    if (isGlobal) {
+      // Global Cursor MCP configuration
+      return path.join(homedir(), '.cursor', 'mcp.json');
+    } else {
+      // Workspace-specific MCP configuration
+      const workspaceDir = process.cwd();
+      return path.join(workspaceDir, '.cursor', 'mcp.json');
+    }
+  }
+
+  async setupWindsurf(options: {
+    extended?: boolean;
+    consumerProject?: boolean;
+    includeOnly?: string;
+    includeAdditional?: string;
+    exclude?: string;
+    isGlobal: boolean;
+  }): Promise<void> {
+    const { isGlobal } = options;
+
+    // Determine mcp.json path
+    const mcpConfigPath = this.getWindsurfSettingsPath(isGlobal);
+
+    // Ensure directory exists
+    await fs.ensureDir(path.dirname(mcpConfigPath));
+
+    // Read existing MCP configuration or create empty object
+    const mcpConfig = await this.readJsonFile(mcpConfigPath);
+
+    // Build MCP server args
+    const args = this.buildMcpServerArgs(options);
+
+    // Create or update MCP configuration for Windsurf
+    if (!mcpConfig.mcpServers) {
+      mcpConfig.mcpServers = {};
+    }
+
+    mcpConfig.mcpServers.bit = {
+      type: 'stdio',
+      command: 'bit',
+      args: args,
+    };
+
+    // Write updated MCP configuration
+    await fs.writeFile(mcpConfigPath, JSON.stringify(mcpConfig, null, 2));
+  }
+
+  getWindsurfSettingsPath(isGlobal: boolean): string {
+    if (isGlobal) {
+      // Global Windsurf MCP configuration
+      return path.join(homedir(), '.windsurf', 'mcp.json');
+    } else {
+      // Workspace-specific MCP configuration
+      const workspaceDir = process.cwd();
+      return path.join(workspaceDir, '.windsurf', 'mcp.json');
+    }
+  }
+
+  async setupEditor(
+    editor: string,
+    options: {
+      extended?: boolean;
+      consumerProject?: boolean;
+      includeOnly?: string;
+      includeAdditional?: string;
+      exclude?: string;
+      isGlobal: boolean;
+    }
+  ): Promise<void> {
+    const supportedEditors = ['vscode', 'cursor', 'windsurf'];
+    const editorLower = editor.toLowerCase();
+
+    if (!supportedEditors.includes(editorLower)) {
+      throw new Error(`Editor "${editor}" is not supported yet. Currently supported: ${supportedEditors.join(', ')}`);
+    }
+
+    if (editorLower === 'vscode') {
+      await this.setupVSCode(options);
+    } else if (editorLower === 'cursor') {
+      await this.setupCursor(options);
+    } else if (editorLower === 'windsurf') {
+      await this.setupWindsurf(options);
     }
   }
